Refactoring PlanetsRenderer
Resolves #172
